requirements.txt
pymongo
table2ascii
discord
python-dotenv
